% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixConstruction.R
\name{MatrixCreation}
\alias{MatrixCreation}
\title{Matrix Construction}
\usage{
MatrixCreation(
  District,
  StandID,
  OSIZE,
  ODEN,
  OSPP1,
  OSPP2,
  USIZE,
  UDEN,
  USPP1,
  USPP2,
  PCT,
  HighElevation = FALSE,
  ACRES = NA
)
}
\arguments{
\item{District}{The SILC District where the stand is located.}

\item{StandID}{A unique identifier for the stand.}

\item{OSIZE}{The overstory stem size class (1 = regenerating, 2 = pre-commercial, 3 = pole, 4 = sawlog).}

\item{ODEN}{The overstory density. (A, B, C, or D)}

\item{OSPP1}{The overstory species 1. Valid species codes are in details}

\item{OSPP2}{The overstory species 2 (if present). Valid species codes are the same as for OSPP1.}

\item{USIZE}{The understory stem size class (1 = regenerating, 2 = pre-commercial, 3 = pole, 4 = sawlog).}

\item{UDEN}{The understory density (A, B, C, or D)}

\item{USPP1}{The understory species 1. Valid species codes are the same as for OSPP1.}

\item{USPP2}{The understory species 2 (if present). Valid species codes are the same as for OSPP1.}

\item{PCT}{The year the stand received pre-commercial thinning treatment.}

\item{HighElevation}{A logical indicating whether the stand is at a high elevation.}

\item{ACRES}{The number of acres in the stand.}
}
\value{
A matrix containing the stand data in a format suitable for use in forest growth and yield models.

Matrix Assignment for Stands Using Overstory and Understory Observations
}
\description{
This function creates a Matrix ID using the 2023 Matrix
Assignment Logic utilizing the dominate condition for matrix labels.
}
\details{
MatrixCreation Function

This function converts stand data from the SILC format into a matrix for use in forest growth and yield models.

This function creates a Matrix ID using the 2023 Matrix
Assignment Logic utilizing the dominate condition for matrix labels

This includes OS strata for RY, AW, and AE, the elimination Of CS and CH strata
replaced by a unified C strata, and the elimination of C strata in RY where they
are lumped with the other OS species.

The Size classes can be 1 (regenerating), 2 (pre-commercial), 3 (pole), or 4 (sawlog). The valid species codes for OSPP1 and OSPP2 (overstory species) and USPP1 and USPP2 (understory species) are as follows:
- "BA": Black Ash
- "BC": Black Cherry
- "BE": Beech
- "BI": Bigtooth aspen
- "BP": Black spruce
- "CE": Cedar
- "BF": Balsam fir
- "BS": Black spruce
- "HE": Hemlock
- "IH": Intolerant Hardwoods
- "MM": Mountain maple
- "NS": Norway Spruce
- "OH": Other Hardwoods
- "OS": Other softwoods
- "PO": Aspen
- "RP": Red pine
- "RM": Red maple
- "RO": Red Oak
- "RS": Red Spruce
- "SF": Spruce-Fir
- "SM": Sugar maple
- "TA": Tamarack
- "TH": Tolerant Hardwood
- "WA": White ash
- "WB": White birch
- "WP": White pine
- "WS": White spruce
- "YB": Yellow birch
}
\examples{
MatrixCreation("RY", "Stand1", 4, "B", "BF", "OS", 2, "C", "BA", "BS", 1999, TRUE, 29)

}
\concept{SILC Functions}
